Tentative conversion manifest v1 to v2
The problem of this approach is that the
manifest v1 does not contain the later DiffIDs
required by the v2 format.

https://github.com/opencontainers/image-spec/blob/v1.0.0/config.md#layer-diffid

The DiffID of a layer is the sha256 checksum of the uncompressed tar.

Apparently the only way to infer it, it'd consist to download and
uncompress each layer, which it will require too much resources and time.