Use tini to propagate signals and reap zombie processes
This commit add the use of tini [1] in order to correct
dispatch signal to children process and handle correctly
terminated processes.

This behavior is expected to be performed by the PID 1 process.
However the 'sh' process does not provide this capabilities [2].

The trick consists using the 'sh' as entry point, then use 'exec tini'
so that the current shell is replaced by the 'tini' process.

1. https://github.com/krallin/tini
2. https://blog.phusion.nl/2015/01/20/docker-and-the-pid-1-zombie-reaping-problem/