Add support for Kubernetes based image build
This commit add the support for building the container
using Kaniko running into a K8s pod.

This has been done refactoring the build logic via an abstract
class named 'BuildStrategy' that comes with two implementations:
- DockerBuildStrategy
- KubeBuildStrategy

The latter is enabled when in the config is provided the option
'wave.build.k8s'. The KubeBuildStrategy connects to a K8s cluster
taking the configuration either via the local kube/config file
or the service account configured in the pod itself, when Wave runs
in a K8s cluster. See the 'K8sClient' implementation for further
details.

Signed-off-by: Paolo Di Tommaso <paolo.ditommaso@gmail.com>