Fix download failed issue caused by relative redirect path
This commit fixes an issue when downloading blobs resulting
in the following error message returned by the docker client:

error pulling image configuration: download failed after attempts=1: unknown blob

The error is caused by a redirect response returned by the registry
in which the location header contains a relative path instead of a
fully qualified URI. Since the location path is resolved by the client
against the Wave host name instead of the source registry host name
this was causing the above error.

This patch solves the issue resolving the redirect location against the
target registry before returning it to the docker client.

Signed-off-by: Paolo Di Tommaso <paolo.ditommaso@gmail.com>