Fix Invalid size in container manifest
This commit fixes the size attribute in the container image
manifest that was reporting the original size instead of the
size of the config of resulting after the augmentation process.

The invalid value was not causing any problem with Docker and
Podman container runtimes, however is reported as an error by
containerd (and recent version of kubernetes which uses the latter
as container runtime).

Signed-off-by: Paolo Di Tommaso <paolo.ditommaso@gmail.com>