Improve response handling and binary download
This commit improves the handling of HTTP requests made via
the Java HttpClient to make sure the response body is *always*
used and the associated InputStream close when avaiable.

This is required otherwise it can result in a never released
SelectorManager thread associated with HttpClient object and
corresponsing memory leak. Read more here https://bugs.openjdk.org/browse/JDK-8308364

It also includes the HTTP status code 303 as a redirection code.

Signed-off-by: Paolo Di Tommaso <paolo.ditommaso@gmail.com>