Fix Container 302 redirect response with body
This commit fixes the handling of redirect response made by
some registry e.g. Quay.io which can provide a repsonse body
other than the Location header directive.

In this situation Wave was discarding the response body causing
the client to not close correcly the connection because it was
awaiting the payload as declared by the Length header.

Signed-off-by: Paolo Di Tommaso <paolo.ditommaso@gmail.com>