Fix Container 302 redirect response with body (2nd version)
This commit fixes the handling of redirect response made by
some registry e.g. Quay.io which can provide a repsonse body
other than the Location header directive.

The patch consists of just overring the Content-Length header
to zero, instead of using the length returned by the original
request and reading the response body as done in the commit
1a90cbc2.

Signed-off-by: Paolo Di Tommaso <paolo.ditommaso@gmail.com>