Add support for blob caching (#359)
This commit adds the ability to cache into an S3-compatible bucket
container layer blobs to prevent multiple downloads from the remote
registry of the same multiple times. Moreover, it offloads the serving of
the blob payloads to the object storage, creating a pre-signed HTTP URL
for the layer blobs to which the Docker client is redirected.

Signed-off-by: Paolo Di Tommaso <paolo.ditommaso@gmail.com>
Signed-off-by: munishchouhan <hrma017@gmail.com>
Co-authored-by: Munish Chouhan <hrma017@gmail.com>
Signed-off-by: Dr Marco Claudio De La Pierre <marco.delapierre@gmail.com>