Add support for Container mirroring (#646)
This commit adds the support for container mirror functionality. 

Container mirror allows creating a copy of a container repository to a target container registry 
keeping the unchanged the  repository name and the container checksums, and therefore 
creating an identical copy to another registry. 

Container mirror is an extension of the Wave build-on-demand feature.  For example 

```
wave -i docker.io/my/foo:bar --mirror-to quay.io
```

it will create a copy of the specified repository using the target image name `quay.io/my/foo:bar`

Signed-off-by: Paolo Di Tommaso <paolo.ditommaso@gmail.com>